restore QueryStep after merging
